servers/faucet: rename binary to faucetd
