+++ servers/faucet/.gitignore
@@ -1,1 +1,1 @@
-faucet-backend
+faucetd
