+++ servers/faucet/Dockerfile
@@ -4,12 +4,12 @@ WORKDIR /build
 
 ADD . .
 
-RUN go build -v -o ./faucet-backend .
+RUN go build -v -o ./faucetd .
 
 FROM alpine AS final
 
 RUN apk add curl
 
-COPY --from=builder /build/faucet-backend /usr/bin/
+COPY --from=builder /build/faucetd /usr/bin/
 
-CMD ["faucet-backend"]
+CMD ["faucetd"]
