+++ servers/faucet/README.md
@@ -1,4 +1,4 @@
-# `faucet-backend`
+# `faucetd`
 
 A simple Go-based REST API that powers the Drivechain faucet at
 [drivechain.live](https://drivechain.live).
