+++ servers/faucet/buf.yaml
@@ -2,7 +2,7 @@
 version: v2
 modules:
   - path: proto
-    name: buf.build/layertwo-labs/faucet-backend
+    name: buf.build/layertwo-labs/faucetd
 deps:
   - buf.build/bitcoin/bitcoind
   - buf.build/bufbuild/protovalidate
