+++ servers/faucet/gen/faucet/v1/faucetv1connect/faucet.connect.go
@@ -8,7 +8,7 @@ import (
 	connect "connectrpc.com/connect"
 	context "context"
 	errors "errors"
-	v1 "github.com/LayerTwo-Labs/sidesail/faucet-backend/gen/faucet/v1"
+	v1 "github.com/LayerTwo-Labs/sidesail/servers/faucet/gen/faucet/v1"
 	http "net/http"
 	strings "strings"
 )
