+++ servers/faucet/server/server.go
@@ -14,10 +14,10 @@ import (
 
 	"connectrpc.com/connect"
 	"connectrpc.com/grpcreflect"
-	api_faucet "github.com/LayerTwo-Labs/sidesail/faucet-backend/api/faucet"
-	"github.com/LayerTwo-Labs/sidesail/faucet-backend/api/faucet/validation"
-	"github.com/LayerTwo-Labs/sidesail/faucet-backend/gen/faucet/v1/faucetv1connect"
-	faucet_ip "github.com/LayerTwo-Labs/sidesail/faucet-backend/ip"
+	api_faucet "github.com/LayerTwo-Labs/sidesail/servers/faucet/api/faucet"
+	"github.com/LayerTwo-Labs/sidesail/servers/faucet/api/faucet/validation"
+	"github.com/LayerTwo-Labs/sidesail/servers/faucet/gen/faucet/v1/faucetv1connect"
+	faucet_ip "github.com/LayerTwo-Labs/sidesail/servers/faucet/ip"
 	coreproxy "github.com/barebitcoin/btc-buf/server"
 	"github.com/rs/cors"
 	"github.com/rs/zerolog"
